move the lm dataset configs to data.text, add shardabledataset to exports
